tests/wpt/web-platform-tests/css/css-anchor-position/anchor-position-multicol-nested-001.html

<!DOCTYPE html>
<title>Anchors in nested multicol</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-pos">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.cb {
  position: relative;
  border-style: solid;
  border-width: 3px 2px 4px 1px;
  padding: 3px 2px 4px 1px;
}
.columns {
  column-count: 4;
  column-fill: auto;
  column-gap: 10px;
  column-width: 150px;
  width: 630px;
  height: 100px;
}
.columns .columns {
  column-count: 2;
  column-width: 60px;
  width: 130px;
  height: 200px;
}
.spacer {
  background: pink;
}
.anchor1 {
  anchor-name: --a1;
  margin-left: 10px;
  width: 20px;
  height: 150px;
  background: blue;
}
.target {
  position: absolute;
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
  background: lime;
  opacity: .3;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <p>The green box should be a union of blue boxes,
    and the purple box should be at the right-bottom of the green box.
  </p>
  <div class="cb">
    <div class="spacer" style="height: 10px"></div>
    <div class="columns">
      <div class="cb">
        <div class="spacer" style="height: 110px"></div>
        <div class="columns">
          <div class="cb">
            <div class="spacer" style="height: 60px"></div>
            <div class="anchor1"></div>
            <div class="spacer" style="height: 80px"></div>
            <div class="target"
                 data-offset-x=11 data-offset-y=-19
                 data-expected-width=180 data-expected-height=100></div>
          </div>
        </div>
        <div class="target"
             data-offset-x=13 data-offset-y=97
             data-expected-width=180 data-expected-height=100></div>
      </div>
    </div>
    <div class="target"
         data-offset-x=175 data-offset-y=13
         data-expected-width=180 data-expected-height=100></div>
  </div>
</body>
